Add X11 to default path for user apps
Use launchctl to set the default PATH for users (takes effect after reboot). Using the "old" path to minimize potential weirds with unexpected custom paths. The ssh_config should no longer be necessary when this is set, however the sshd_config would still be needed. (The launchd setting could alsö be applied to the system context, but I'm more weary of messing with system processes.)